05_Projects/Student_Feedback_System/requirements.txt
gradio
Pyrebase4
python-dotenv
pandas
